+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 Recommended Gamevault Server Version: `v12.1.0`
 ### Changes
 - Added Encrypted (password protected) archive support. Including a default password setting.
-- Added the possibility to pause/resume downloads. Paused downloads are restored after the restart.
+- Added the possibility to pause/resume downloads. Paused downloads are restored after the restart. Also pause is triggered on error.
+- Added auto retry after download failed.
+- Show system notifications for download/extraction
 - Bug fix: Center game view background image on window resize
 - Bug fix: Tag names that are too long could not be selected in the filters
 - Bug fix: Displayed download speed has adjusted very slowly with larger fluctuations